+++ calibrate_colors.cpp
